+++ package.json
@@ -229,10 +229,10 @@
     }
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:node": "npm run -s clean:node && webpack --env production",
-    "build:preview": "npm run -s clean:preview && webpack --env production --config webpack.preview.config.js",
-    "build:web": "npm run -s clean:web && webpack --env production --config webpack.web.config.js",
+    "build": "npm run -s clean && webpack --env production",
+    "build:node": "npm run -s clean:node && webpack --env production --env target=node",
+    "build:preview": "npm run -s clean:preview && webpack --env production --env target=preview",
+    "build:web": "npm run -s clean:web && webpack --env production --env target=web",
     "clean": "run-p clean:*",
     "clean:node": "rimraf lib",
     "clean:preview": "rimraf preview",
@@ -250,21 +250,14 @@
     "test:unit:coverage": "jest --coverage",
     "version": "curl https://raw.githubusercontent.com/marp-team/actions/v1/lib/scripts/version.js | node && git add -A CHANGELOG.md",
     "vsce:publish": "vsce publish --no-yarn",
-    "vscode:prepublish": "run-s preversion build:*",
-    "watch": "run-p _watch:*",
-    "watch:node": "run-p _watch:node _watch:preview",
-    "watch:web": "run-p _watch:web _watch:preview",
-    "_watch:node": "webpack --watch",
-    "_watch:preview": "webpack --watch --config webpack.preview.config.js",
-    "_watch:web": "webpack --watch --config webpack.web.config.js"
+    "vscode:prepublish": "run-s preversion build",
+    "watch": "webpack --watch",
+    "watch:node": "webpack --watch --env target=node,preview",
+    "watch:web": "webpack --watch --env target=web,preview"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.15.4",
     "@marp-team/marp-core": "^2.1.1",
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@rollup/plugin-typescript": "^8.2.5",
     "@types/express": "^4.17.13",
     "@types/jest": "^27.0.1",
     "@types/lodash.debounce": "^4.0.6",
@@ -273,8 +266,6 @@
     "@typescript-eslint/eslint-plugin": "^4.30.0",
     "@typescript-eslint/parser": "^4.30.0",
     "abort-controller": "^3.0.0",
-    "bufferutil": "^4.0.3",
-    "builtin-modules": "^3.2.0",
     "codecov": "^3.8.3",
     "dedent": "^0.7.0",
     "esbuild": "^0.12.25",
@@ -301,8 +292,6 @@
     "remark-math": "^5.1.0",
     "remark-parse": "^10.0.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.56.3",
-    "rollup-plugin-terser": "^7.0.2",
     "stylelint": "^13.13.1",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-standard": "^22.0.0",
@@ -312,7 +301,6 @@
     "typescript": "~4.3.5",
     "unified": "^10.1.0",
     "unist-util-visit": "^4.0.0",
-    "utf-8-validate": "^5.0.5",
     "vsce": "^1.96.3",
     "webpack": "^5.52.0",
     "webpack-cli": "^4.8.0",
